refactor: shift util functions to shared folder to prevent code duplication
