+++ frontend/src/utils/date.ts
@@ -160,39 +160,3 @@ export const transformCheckedBoxesValueToInvalidDays = (
   const validDaysSet = new Set(validDays)
   return ALL_INVALID_DAYS_ARR.filter((day) => !validDaysSet.has(day))
 }
-
-const DAY_TO_NUMBER_MAP: Record<InvalidDaysOptions, number> = {
-  [InvalidDaysOptions.Monday]: 1,
-  [InvalidDaysOptions.Tuesday]: 2,
-  [InvalidDaysOptions.Wednesday]: 3,
-  [InvalidDaysOptions.Thursday]: 4,
-  [InvalidDaysOptions.Friday]: 5,
-  [InvalidDaysOptions.Saturday]: 6,
-  [InvalidDaysOptions.Sunday]: 7,
-}
-
-/**
- * Convert the days of the week in the invalidDays array
- * to a number array representing the number representation
- * of the corresponding day of the week
- */
-export const convertInvalidDaysOfTheWeekToNumberSet = (
-  invalidDays: InvalidDaysOptions[],
-): Set<number> => {
-  if (invalidDays.length === 0) {
-    return new Set()
-  }
-
-  return new Set(invalidDays.map((invalidDay) => DAY_TO_NUMBER_MAP[invalidDay]))
-}
-
-export const isDateAnInvalidDay = (
-  date: Date,
-  invalidDays: InvalidDaysOptions[],
-): boolean => {
-  const invalidDaySet = convertInvalidDaysOfTheWeekToNumberSet(invalidDays)
-
-  const dayNumberFormat = parseInt(format(date, 'i'))
-
-  return invalidDaySet.has(dayNumberFormat)
-}
