+++ src/app/utils/field-validation/validators/dateValidator.ts
@@ -3,10 +3,8 @@ import { chain, left, right } from 'fp-ts/lib/Either'
 import { flow } from 'fp-ts/lib/function'
 import moment from 'moment-timezone'
 
-import {
-  DateSelectedValidation,
-  InvalidDaysOptions,
-} from '../../../../../shared/types'
+import { DateSelectedValidation } from '../../../../../shared/types'
+import { convertInvalidDaysOfTheWeekToNumberSet } from '../../../../../shared/utils/date-validation'
 import {
   IDateFieldSchema,
   OmitUnusedValidatorProps,
@@ -108,31 +106,6 @@ const makeDateValidator: DateValidatorConstructor = (dateField) => {
     default:
       return right
   }
-}
-
-const DAY_TO_NUMBER_MAP: Record<InvalidDaysOptions, number> = {
-  [InvalidDaysOptions.Monday]: 1,
-  [InvalidDaysOptions.Tuesday]: 2,
-  [InvalidDaysOptions.Wednesday]: 3,
-  [InvalidDaysOptions.Thursday]: 4,
-  [InvalidDaysOptions.Friday]: 5,
-  [InvalidDaysOptions.Saturday]: 6,
-  [InvalidDaysOptions.Sunday]: 7,
-}
-
-/**
- * Convert the days of the week in the invalidDays array
- * to a number array representing the number representation
- * of the corresponding day of the week
- */
-const convertInvalidDaysOfTheWeekToNumberSet = (
-  invalidDays: InvalidDaysOptions[],
-): Set<number> => {
-  if (invalidDays.length === 0) {
-    return new Set()
-  }
-
-  return new Set(invalidDays.map((invalidDay) => DAY_TO_NUMBER_MAP[invalidDay]))
 }
 
 /**
